app/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.5
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environ==0.9.0
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
whitenoise==6.4.0
